cc, gpu: Widen PaintCacheId and paint cache deletion IPC to uint64_t
Skia is widening SkPath generation IDs from 32-bit to 64-bit. This CL
prepares Chromium to consume 64-bit IDs by widening PaintCacheId and
its associated serialization and GPU command buffer purging pipeline.

Bug: 562038679
Bug: 556268833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8429709
Commit-Queue: Thomas Smith <thomsmit@google.com>
Cr-Commit-Position: refs/heads/main@{#1703606}

+++ cc/paint/paint_cache.h
@@ -38,7 +38,7 @@ namespace cc {
 // since it avoids the need for cross-process ref-counting required by the
 // TransferCache.
 
-using PaintCacheId = uint32_t;
+using PaintCacheId = uint64_t;
 using PaintCacheIds = std::vector<PaintCacheId>;
 enum class PaintCacheDataType : uint32_t {
   kPath,
